ingestion_service/producer/requirements.txt
fastapi==0.110.2
#richlogging
rich == 13.7.1
#python-json-logger
python-json-logger == 2.0.7
pydantic-settings==2.2.1
python-multipart==0.0.9
# to handle multiple user logging
asgi-correlation-id ==4.3.1
#logtailhanlder
logtail-python ==0.2.10
# to handle type annotation 'int | None' future
eval-type-backport == 0.1.3
pika==1.3.2
pkonfig==2.0.0
psycopg==3.1.18
python-jose==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
asyncpg==0.29.0